e2,kpm: fix RLC metric numeric precision

+++ lib/e2/e2sm/e2sm_kpm/e2sm_kpm_du_meas_provider_impl.h
@@ -79,10 +79,10 @@ private:
   bool check_e2sm_kpm_metrics_definitions(span<const e2sm_kpm_metric_t> metrics_defs);
 
   // Helper functions.
-  float bytes_to_kbits(float value);
-  bool  handle_no_meas_data_available(const std::vector<asn1::e2sm::ue_id_c>&        ues,
-                                      std::vector<asn1::e2sm::meas_record_item_c>&   items,
-                                      asn1::e2sm::meas_record_item_c::types::options value_type);
+  double bytes_to_kbits(double value);
+  bool   handle_no_meas_data_available(const std::vector<asn1::e2sm::ue_id_c>&        ues,
+                                       std::vector<asn1::e2sm::meas_record_item_c>&   items,
+                                       asn1::e2sm::meas_record_item_c::types::options value_type);
 
   // Measurement getter functions.
   metric_meas_getter_func_t get_cqi;
